all: Add US_902_928 band definition
