+++ pkg/band/all.go
@@ -97,6 +97,15 @@ var All = map[string]map[ttnpb.PHYVersion]Band{
 		ttnpb.RP001_V1_1_REV_A:   RU_864_870_RP1_V1_1_Rev_A,
 		ttnpb.RP001_V1_1_REV_B:   RU_864_870_RP1_V1_1_Rev_B,
 	},
+	US_902_928: {
+		ttnpb.TS001_V1_0:         US_902_928_TS1_V1_0,
+		ttnpb.TS001_V1_0_1:       US_902_928_TS1_V1_0_1,
+		ttnpb.RP001_V1_0_2:       US_902_928_RP1_V1_0_2,
+		ttnpb.RP001_V1_0_2_REV_B: US_902_928_RP1_V1_0_2_Rev_B,
+		ttnpb.RP001_V1_0_3_REV_A: US_902_928_RP1_V1_0_3_Rev_A,
+		ttnpb.RP001_V1_1_REV_A:   US_902_928_RP1_V1_1_Rev_A,
+		ttnpb.RP001_V1_1_REV_B:   US_902_928_RP1_V1_1_Rev_B,
+	},
 }
 
 // Get returns the band if it was found, and returns an error otherwise.
